Changed all routes to use app.router and split routes into separate files.
